FIX imports for moved post components
