+++ client/src/components/SearchResults.tsx
@@ -3,7 +3,7 @@ import axios from 'axios';
 import { useParams } from 'react-router-dom';
 import Box from '@mui/material/Box';
 import SearchComponent from './Search';
-import Post from './Post';
+import Post from './posts/post card/Post';
 import { PostWithRelations } from '../../../types';
 
 export default function SearchResults() {
